Scale cells to 3 in all envs
- For environments with only two cells, the apps will pile up on
whichever cell rolls first during a rolling deploy. This is because they
evacuate the 1st cell onto the 2nd, and then evacuate back to the 1st
cell every time. Over time this will result in a de-facto single cell
environment because diego will never schedule anything for the full
cell.

Signed-off-by: Greg Cobb <gcobb@pivotal.io>