Add use-pre-spec-refactor-properties.yml ops file to kiki
- Kiki deploys the last cf release version of capi, and then rolls the
migration forward to the current state in order to check backwards
compatibility. In our recent refactor of our configs, we got rid of a
bunch of unnecessary properties. We also removed them from the
cf-deployment manifest. This means that Kiki was trying to deploy the
old version of capi with the updated deployment manifest, causing the
deploy to fail due to the lack of config properties.
- Kiki is not concerned with how it gets deployed, so long as it tests
the backwards compatibility of the migrations and db.
- Once we cut a new CF release with the configuration refactor included,
we will no longer need to use the `pre-spec-refactor-properties.yml` ops
file.

Signed-off-by: Elena Sharma <esharma@pivotal.io>