Stop eval-ing bbl print-env in DRATS task
This was exporting the BOSH_ALL_PROXY variable which was instructing
the bbr CLI to attempt to use a SOCKS proxy. This is a new feature for
bbr and it currently appears to have some issues around concurrent `bosh
ssh` connections. For now we will avoid using it until the BBR team is
running DRATS with it in their own pipelines.

See: https://www.pivotaltracker.com/n/projects/1662777/stories/159231402

We actually did not need to eval bbl print-env in this task so removing
it will prevent the BOSH_ALL_PROXY variable from being exported.

[#159407486]

Signed-off-by: Tim Downey <tdowney@pivotal.io>