Use cf-deployment-concourse-tasks/run-cats to run CATS
The cats-concourse-task task that we were using to run
acceptance tests in our environment was deprecated in March
and is packaged with an older version of the CF CLI

A recent change to CATS requires support for `cf tail` so
we need to use the recommended run-cats task from
cf-deployment-concourse-tasks

Additionally, the version tag filtering on the
cf-deployment-concourse-tasks resource was removed while
troubleshooting this issue so it started pulling in newer versions
of that resource which are incompatible with our current
environments. Since there is not an easy way to clear fetched
versions from a resource we had to rename it.

See: https://github.com/concourse/concourse/issues/145

Signed-off-by: Eli Wrenn <ewrenn@pivotal.io>