BOSH Lites use SSDs
GCP SSDs are 22× faster than regular disks (`pd-ssd` vs. `pd-standard`);
Source: [Pivotal Engineering
Journal](http://engineering.pivotal.io/post/gobonniego_results/)

The BOSH Lite's speed is heavily dependent on the speed of its disk,
so we're speeding it up.

SSDs are 4× as expensive, and a given BOSH Lite has approximately 96GB
storage, which works out to $16.32/month. The price of the n1-standard-4
instances we use are $97.09/month, so the cost of the disk is 1/6 the
cost of the instance.

[finishes #166994226]

Authored-by: Brian Cunnie <bcunnie@pivotal.io>