🐞 Remove terraform's deprecation warnings
fixes:
```
Warning: Interpolation-only expressions are deprecated

  on bosh-lite.tf line 53, in resource "google_compute_firewall" "default":
  53:   network = "${google_compute_network.default.name}"

Terraform 0.11 and earlier required all non-constant expressions to be
provided via interpolation syntax, but this pattern is now deprecated. To
silence this warning, remove the "${ sequence from the start and the }"
sequence from the end of this expression, leaving just the inner expression.

Template interpolation syntax is still used to construct strings from
expressions when the template includes multiple interpolation sequences or a
mixture of literal strings and interpolations. This deprecation applies only
to templates that consist entirely of a single interpolation sequence.
```
```
Warning: Quoted references are deprecated

  on bosh-lite.tf line 74, in resource "google_dns_record_set" "default":
  74:   provider = "google.dns"

In this context, references are expected literally rather than in quotes.
Terraform 0.11 and earlier required quotes, but quoted references are now
deprecated and will be removed in a future version of Terraform. Remove the
quotes surrounding this reference to silence this warning.
```

[#169793167]

Co-authored-by: Reid Mitchell <rmitchell@pivotal.io>
Co-authored-by: Brian Cunnie <bcunnie@pivotal.io>