Use handmade kubectl task to apply kpack
The reason we converted from the old way of using the kubectl concourse
resource is because it was difficult to obtain the kubeconfig file. In
katniss, the kubectl resource followed a task which extracted the
kubeconfig file out of a bbl directory. Since we do not use bbl for
samus (a k8s only environment), it was simpler to create a bash script
task to apply yaml using kubectl. That allows us to use the envrc file
within the capi-ci-private/samus directory, which generates the
kubeconfig file.

[#171063670]

Co-authored-by: Chris Tarazi <ctarazi@pivotal.io>
Co-authored-by: Sannidhi Jalukar <sjalukar@pivotal.io>