Remove tag filter on cf-deployment-concourse-tasks
This was previously introduced to avoid pulling in changes that hadn't
been released by rel-int. However, there has not been a release recently
and we are behind in the version of the cf CLI. We think that removing
this filter should be a safe change given that there is not a lot of
development in cf-deployment-concourse-tasks right now, but once a new
release is cut with a CLI >7.2, we can bring this filter back if needed.

[#175042021](https://www.pivotaltracker.com/story/show/175042021)

Co-authored-by: Sarah Weinstein <sweinstein@pivotal.io>
Co-authored-by: Belinda Liu <bliu@pivotal.io>