Switch Elsa to cf-deployment-concourse-tasks to deploy
Elsa was previously configured to use the bosh-deployment-resource to
deploy. This is no longer a pattern we want to follow as all other
environments have migrated to using cf-deployment-concourse-tasks to
deploy instead.

This commit uses tasks to collect ops files and vars files that `elsa`
will need before deploying. It also removes the stemcell-upload tasks,
as stemcells are uploaded by the cf-deployment-concourse-tasks deploy
task.

[#176753798](https://www.pivotaltracker.com/story/show/176753798)

Co-authored-by: Belinda Liu <bliu@pivotal.io>
Co-authored-by: Weyman Fung <weymanf@vmware.com>