Clean up unnecessary extract-bbl-environment tasks
This task is a vestige from when environments were deployed with the
bosh-deployment-resource, which required a source file that is now
encompassed in a bbl-state file. The output of this task is still used
in some other tasks (such as drats and blobstore-benchmarks), but is no
longer necessary for certain jobs.

[#176753798](https://www.pivotaltracker.com/story/show/176753798)

Authored-by: Belinda Liu <bliu@pivotal.io>