Make bridge unit tests work with go mod enabled
A recent version of locket breaks when go modules are disabled.
However, locket cannot currently be built as a standalone library
because its go.mod file is not checked in to the locket repo - it's
checked under the `src` directory in the `diego-release` repo.

The locket maintainers are working to address this issue, but we need a
temporary solution to fix the tps-watcher tests.

The test job will now build the locket binary and place it on the $PATH
(which the tps-watcher tests now expect).

We also had to run the tps-watcher tests and the cc-uploader tests with
two separate ginkgo invocations because they need to be run from the
same folder as the go.mod file.

Co-authored-by: Tom Viehman <tviehman@pivotal.io>
Co-authored-by: Weyman Fung <weymanf@vmware.com>